% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoB-data.r
\docType{data}
\name{ChoB}
\alias{ChoB}
\title{A dataset to check the sensitivity of the violation of essential tau-equivalence}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\usage{
data(ChoB)
}
\description{
A hypothetical data consisting of three items. Item 1 has a variance of 1,
 item 2 is a copy of item 1, and item 3 is item 1 divided by 2. This is
 the data presented by Cho (in press) to show that coefficient alpha is
 sensitive to the violation of the assumption of being essentially
 tau-equivalent.
}
\references{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
 omega: A comment on Sijtsma and Pfadt. Psychometrika.
 @author Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
