% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.r
\name{cr}
\alias{cr}
\title{Obtain congeneric reliability (Unidimensional CFA reliability)}
\usage{
cr(data)
}
\arguments{
\item{data}{a dataframe or a matrix (unidimensional)}
}
\value{
congeneric reliability coefficient
}
\description{
Congeneric reliability is a reliability coefficient derived from
unidimensional confirmatory factor analysis (CFA).
}
\details{
Dependency: This function requires the lavaan package installed.

Features: Congeneric reliability is a single-dimensional reliability
coefficient based on a congeneric model. The congeneric model is a model that
allows the length, discrimination, or importance of items to be different,
and is the least restrictive model among the models derived from the
classical test theory. Congeneric reliability uses an optimization technique
called maximum likelihood to estimate the 'length' of an item.

Name: Congeneric reliability is called by a variety of names, general users
usually call it composite reliability, and reliability researchers often call
it omega. One of the reasons for this confusion is that studies that first
proposed this coefficient (Joreskog 1971) did not give this formula a name
 (Cho 2016). Joreskog (1971) proposed a matrix-form formula, and the
 commonly known non-matrix formula appears in Werts et al. (1974).

Frequency of use: Congeneric reliability is the second most commonly used
reliability coefficient after coefficient alpha (Cho 2016)

Accuracy: Congeneric reliability is the most accurate reliability coefficient
along with the Feldt-Gilmer coefficient (Cho in press)

Computation: This function uses maximum likelihood as estimation,
unstandardized covariance matrix as input, and lavaan package as software.
First, this function estimates the CFA model with the factor's variance fixed
at 1.'F =~ NA*V1 + ...Vk \n F ~~ 1*F' Next, the sum of the obtained factor
loadings is squared (A) and the errors are summed (B). Cr is the value of
A/(A+B).
}
\examples{
cr(Graham1)
}
\references{
Cho, E. (2016). Making reliability reliable: A systematic
approach to reliability coefficients. Organizational Research Methods, 19(4),
651-682.

Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
omega: A comment on Sijtsma and Pfadt. Psychometrika.

Jöreskog, K. G. (1971). Statistical analysis of sets of
congeneric tests. Psychometrika, 36(2), 109-133.

Werts, C. E., Linn, R. L., & Jöreskog, K. G. (1974). Intraclass
reliability estimates: Testing structural assumptions. Educational and
Psychological Measurement, 34, 25-33.
}
\seealso{
[gilmer()] for the Gilmer-Fedlt coefficient

[cc()] for classical congeneric reliability coefficient

[psych::omega()] for a related function of the package psych

[MBESS::ci.reliability()] for a related function of the package MBESS

[Lambda4::omega.tot()] for a related function of the package Lambda4
}
\concept{congenerics}
