% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov.r
\name{get_cov}
\alias{get_cov}
\title{Obtain the covariance matrix}
\usage{
get_cov(data)
}
\arguments{
\item{data}{A dataframe or a matrix}
}
\value{
The covariance matrix
}
\description{
If the input data is a square matrix, it is converted into a matrix,
otherwise the covariance matrix is obtained.
The covariance matrix uses the cov function of base,
and the option for processing missing values is 'pairwise.complete.obs'.
}
