% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu0.r
\name{mu0}
\alias{mu0}
\title{Obtain Ten Berge and Zegers' (1978) mu0}
\usage{
mu0(data)
}
\arguments{
\item{data}{a dataframe or a matrix (unidimensional)}
}
\description{
Obtain Ten Berge and Zegers' (1978) mu0. mu0 equals coefficient alpha.
}
\references{
Ten Berge, J. M. F., & Zegers, F. E. (1978). A series of lower
bounds to the reliability of a test. Psychometrika, 43(4), 575-579.
}
\seealso{
[alpha()]
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
