% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu4.R
\name{mu4}
\alias{mu4}
\title{Obtain Ten Berge and Socan's (2004) mu4}
\usage{
mu4(data)
}
\arguments{
\item{data}{a dataframe or a matrix (unidimensional)}
}
\description{
Obtain Ten Berge and Socan's (2004) mu4.
}
\examples{
mu4(Graham1)
}
\references{
Ten Berge, J. M. F., & Zegers, F. E. (1978). A series of lower
bounds to the reliability of a test. Psychometrika, 43(4), 575-579.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
