% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus.R
\name{mus}
\alias{mus}
\title{Obtain Ten Berge and Socan's (2004) mu0 to mu5}
\usage{
mus(data)
}
\arguments{
\item{data}{a dataframe or a matrix (unidimensional)}
}
\value{
mu0 Ten Berge and Socan's mu0 reliability coefficient (coefficient alpha)

mu1 Ten Berge and Socan's mu1 reliability coefficient (Guttman's lambda2)

mu2 Ten Berge and Socan's mu2 reliability coefficient

mu3 Ten Berge and Socan's mu3 reliability coefficient

mu4 Ten Berge and Socan's mu4 reliability coefficient

mu5 Ten Berge and Socan's mu5 reliability coefficient
}
\description{
Obtain Ten Berge and Socan's (2004) mu0, mu1, mu2, mu3, mu4, and mu5.
}
\examples{
mus(Graham1)
}
\references{
Ten Berge, J. M. F., & Zegers, F. E. (1978). A series of lower
bounds to the reliability of a test. Psychometrika, 43(4), 575-579.
}
\seealso{
[psych::tenberg()] for a related function of the package psych
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
