% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reuterberg-data.r
\docType{data}
\name{Reuterberg}
\alias{Reuterberg}
\title{Reuterberg and Gustafsson's (1992) unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\usage{
data(Reuterberg)
}
\description{
Reuterberg and Gustafsson's (1992) actual dataset consisting of five items.
This data is the original 40-item verbal test grouped by eight and reduced to
five items.The CFA reliability coefficient presented by the author is .803.
 There are currently various approaches to FA reliability, and this value
 can be used to check whether these approaches refer to the same thing (Cho
 in press).
}
\references{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
 omega: A comment on Sijtsma and Pfadt. Psychometrika.

Reuterberg, S.-E., & Gustafsson, J.-E. (1992). Confirmatory
factor analysis and reliability: Testing measurement model assumptions.
Educational and Psychological Measurement, 52(4), 795-811.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
