% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unirel.R
\name{unirel}
\alias{unirel}
\title{Obtain various unidimensional reliablity coefficients}
\usage{
unirel(data, Lambda4.include = TRUE, psych.include = TRUE)
}
\arguments{
\item{data}{a dataframe or a matrix (unidimensional)}

\item{Lambda4.include}{Whether to include the reliability coefficients
provided by the package Lambda4 (lambda5, lambda6, max_lambda, lambda4_max,
lambda4_75)}

\item{psych.include}{Whether to include reliability coefficients
(GLB.algebraic, GLB.fa) provided by the package psych}
}
\value{
alpha coefficient alpha (tau-equivalent reliability)

lambda2 Guttman's lambda2

mu2 Ten Berge and Zegers' mu2

mu3 Ten Berge and Zegers' mu3

mu4 Ten Berge and Zegers' mu4

feldt Feldt's classical congeneric reliability coefficient

gilmer the Gilmer-Fedlt coefficient

joreskog unidimensional CFA (congeneric) reliability

hancock Hancock's H

heise Heise-Borhnstedt's Omega

kaiser Kaiser-Caffrey's alpha

lambda5 Guttman's lambda5

lambda5 Guttman's lambda6

max_lambda the maximum among lambda2, lambda5, and lambda6

lambda4_max the maximum among all possible lambda4

lambda4_75 the 75the percentile among all possible lambda4

glb.algebraic the original estimation of the greatest lower bounds

glb.fa an improvement of the greatest lower bounds by the psych package
}
\description{
R packages that provide unidimensional reliability coefficients include
Lambda4 and psych. These are usually separate functions, so comparing
multiple reliability estimates can take a long time. This function shows the
 reliability coefficients newly added in this package together with the
 reliability coefficients provided by Lambda4 and psych. This function can
 provide Table 2 of Cho(in press) immediately.
}
\examples{
unirel(Graham1)
}
\references{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
omega: A comment on Sijtsma and Pfadt. Psychometrika.
}
\seealso{
[Lambda4::lambda5()] for a related function of the package Lambda4

[Lambda4::lambda6()] for a related function of the package Lambda4

[Lambda4::quant.lambda4()] for a related function of the package Lambda4

[psych::GLB.algebraic()] for a related function of the package psych

[psych::GLB.fa()] for a related function of the package psych
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
